MOVEONLY: TestLockPointValidity to txmempool
